\name{localoptions-package}
\alias{localoptions-package}
\alias{localoptions}
\docType{package}
\title{
Loads local options files into options()
}
\description{
	This package contains two functions which loads an configuration
	file with options. These options are stored into R's options
	settings (using options())
}
\details{
\tabular{ll}{
Package: \tab localoptions\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-02-02\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}

Usage: readoptions("configfile.conf")
That's all there is

}
\author{
	Job Spijker (rdev@spkr.nl)

}

\keyword{ package }
\keyword{ options }
